Fix daylight savings bug. Close #72
Ok, so I think this should actually address the problem once and for all. To be honest, I'm not quite sure what the previous code was trying to accomplish.

Previous work:
- #10
- #13
- 8fb2dee782bbb1a3059db27ee654d7553d90a4ee

We are no longer using `TimeZone.current` and `Locale.current` and attempting to figure out how to map the Cupertino Date to another location.

Instead we are using `PST` and `en-US` to specifically target the precise time zone and locale.